+++ src/Refahi.Modules.Hotels.Application.Contracts/Providers/Queries/SearchHotelsQuery.cs
@@ -10,8 +10,8 @@ public sealed record SearchHotelsQuery(
     int? Adults,
     int? Children,
     int? AvailableRooms,
-    int? MinPrice,
-    int? MaxPrice,
+    long? MinPrice,
+    long? MaxPrice,
     int[]? Stars,
     string[]? Accommodations
 ) : IRequest<IEnumerable<HotelSearchResultDto>>;
